Support u32 chart positions and broaden BPM handling

+++ src/stage.h
@@ -149,7 +149,7 @@ typedef struct
 
 typedef struct
 {
-	u16 end; //1/12 steps
+	u32 end; //1/12 steps
 	u16 flag;
 } Section;
 
@@ -166,7 +166,7 @@ typedef struct
 
 typedef struct
 {
-	u16 pos; //1/12 steps
+	u32 pos; //1/12 steps
 	u16 type;
 	u16 is_opponent;
 } Note;
@@ -313,12 +313,12 @@ typedef struct
 		u16 size;
 	} note;
 	
-	u16 last_bpm;
+	u32 last_bpm;
 
 	int timerlength, timermin, timersec, timepassed;
 	
 	fixed_t time_base;
-	u16 step_base;
+	u32 step_base;
 	Section *section_base;
 
 	// Grace period after swaps to avoid false misses
